Create Binary array difference
Given an array of 1's and 0's. Return a positive number if the count of 1's are more than the count of 0's, return a negative number if the number of 1's are less than the number of 0's and return 0 if their counts are same. The first input is the number of elements of the binary array followed by the binary array.