Create MAX_array
Write a program to read an array and find the largest element in the array

While entering the input, the first element should be the n,length of the array. Followed by the n elements (A[i]) of the array. The output should be the largest element in the n element array.