Create Duplicate_count
Write a program to count the number of duplicate elements in an array.

While entering the input, the first element should be the n, length of the array. Followed by the n elements (A[i]) of the array. The output should be the count or duplicate elements in an array